[Decomp][Spell] Cohesion split: extract late crit/mastery/vehicle aura handlers to SpellAuraLate.cpp
Move the handler block that sits between IsCritFromAbilityAura and
IsLastAuraOnHolder (HandleAuraModAllCritChance, HandleAuraStopNaturalManaRegen,
HandleAuraSetVehicleId, HandleAuraMastery, HandleAuraModBlockCritChance,
HandlePreventResurrection, HandleFactionOverride, HandleTriggerLinkedAura) out
of SpellAuras.cpp into a new SpellAuraLate.cpp translation unit.

Pure file-level cohesion split: declarations stay in SpellAuras.h, the
AuraHandler[] table stays in SpellAuras.cpp and resolves the moved member
methods by address at link time. Byte-identical move, no behaviour change.
SpellAuras.cpp 9040 -> 8903.
