+++ src/game/WorldHandlers/SpellAuras.cpp
@@ -6991,143 +6991,6 @@ bool Aura::IsCritFromAbilityAura(Unit* caster, uint32& damage)
     return false;
 }
 
-void Aura::HandleAuraModAllCritChance(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    if (target->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)target)->HandleBaseModValue(CRIT_PERCENTAGE,         FLAT_MOD, float(m_modifier.m_amount), apply);
-    ((Player*)target)->HandleBaseModValue(OFFHAND_CRIT_PERCENTAGE, FLAT_MOD, float(m_modifier.m_amount), apply);
-    ((Player*)target)->HandleBaseModValue(RANGED_CRIT_PERCENTAGE,  FLAT_MOD, float(m_modifier.m_amount), apply);
-
-    // included in Player::UpdateSpellCritChance calculation
-    ((Player*)target)->UpdateAllSpellCritChances();
-}
-
-void Aura::HandleAuraStopNaturalManaRegen(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    GetTarget()->ApplyModFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_REGENERATE_POWER, !apply);
-}
-
-void Aura::HandleAuraSetVehicleId(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    GetTarget()->SetVehicleId(apply ? GetMiscValue() : 0, 0);
-}
-
-void Aura::HandleAuraMastery(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-    if (target->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)target)->UpdateMasteryAuras();
-}
-
-void Aura::HandleAuraModBlockCritChance(bool apply, bool Real)
-{
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->ApplyModUInt32Value(PLAYER_SHIELD_BLOCK_CRIT_PERCENTAGE, m_modifier.m_amount, apply);
-    }
-}
-
-void Aura::HandlePreventResurrection(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-    if (!target || target->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        target->RemoveByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_RELEASE_TIMER);
-    }
-    else if (!target->GetMap()->Instanceable())
-    {
-        target->SetByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_RELEASE_TIMER);
-    }
-}
-
-void Aura::HandleFactionOverride(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-    if (!target || !sFactionTemplateStore.LookupEntry(GetMiscValue()))
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        target->setFaction(GetMiscValue());
-    }
-    else
-    {
-        target->RestoreOriginalFaction();
-    }
-}
-
-
-void Aura::HandleTriggerLinkedAura(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    uint32 linkedSpell = m_spellEffect->EffectTriggerSpell;
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(linkedSpell);
-    if (!spellInfo)
-    {
-        sLog.outError("Aura::HandleTriggerLinkedAura for spell %u effect %u triggering unknown spell id %u", GetSpellProto()->Id, m_effIndex, linkedSpell);
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    if (apply)
-    {
-        // ToDo: handle various cases where base points need to be applied!
-        target->CastSpell(target, spellInfo, true, NULL, this);
-    }
-    else
-    {
-        target->RemoveAurasByCasterSpell(linkedSpell, GetCasterGuid());
-    }
-}
-
 /**
  * @brief Checks whether this aura is the last remaining effect on its holder.
  *
